Changing initial grad value to +1 and changing +=
1. Changed the default value of utop.grad to 1
2. Changing += from backward functions to =

First of all thanks for writing such a brilliant article. This is gold mine for me. I have few suggestions for changes, which I have checked with proof, and even if I am wrong, I am sure I will learn more from you in the process :) 

1. Changing the default value of grad to 1.0
- I believe all the gates we made in this section of neuron example should be able to exist and function independently too. So, in a way I should be able to take the class multiplyGate and use it again to achieve what we did in Strategy #3. And to get a better value, I must do iterations. If the initial value of the grad is zero, then for every iteration I have to reassign utop.grad a value of 1 (like you did s.grad = 1.0)  for every iteration. If the initial value is 1, then I do not need to assign the utop.grad to 1 again and again.
Also by my understanding of backpropogation (completely from your wonderful article), it's a play of chain rule, and since chain rule is multiplication of derivatives, having a default value of 1 will bring no change. So it makes more sense. This way even if we take multiplyGate as a seperate neuron, then the value  of this.u1.value*this.utop.grad will act as it should in a single neuron, thus fulfilling the purpose.

2. Changing += from backward functions to =
Suppose we take a gate of -x^2-y^2
We set the initial values as x= -0.1 and y= - 0.1
Now with the idea of adding gradients, if I do 1000 iterations, there will be a point where I will reach